// src/app/models/balance.model.ts
export interface Balance{
    nombre: string;
    fecha: string;
    mantenimientopagos: string;
    areascomunespgos: string;
    fechagastos: string;
    rubro: string;
    concepto: string;
    monto: string;
    saldomes: string;
    saldocuenta: number;
    importe: number;
    costo: number;
    total: number;
    domicilio: string;
    status: string;
    
}